requirements.txt
gpxpy
moviepy
geopandas
matplotlib
osmx
requests
scipy
tqdm
folium